refactor: 홈 화면 실시간 인기 검색어 게시판 폴더 이동
