+++ src/entities/hotBoard/ui/HotBoardListRow.tsx
