+++ src/entities/hotBoard/ui/MedalRow.tsx
