+++ src/entities/hotBoard/ui/index.ts
@@ -1,0 +1,3 @@
+import HotBoardList from './HotBoardList';
+
+export { HotBoardList };
