+++ src/widgets/hotBoards/ui/index.ts
@@ -1,6 +1,5 @@
 import AdvCarousel from './AdvCarousel';
-import HotBoardList from './HotBoardList';
 import { SortChip, SortChipItem } from './SortChip';
 import Fire from './icons/Fire';
 
-export { AdvCarousel, HotBoardList, SortChip, SortChipItem, Fire };
+export { AdvCarousel, SortChip, SortChipItem, Fire };
